Add field called  into Context record to handle synapse payload
