+++ cli-synapse/src/test/resources/ballerina/HealthCheckApi/main.bal
@@ -5,6 +5,8 @@ public listener http:Listener httpListener = new (8080);
 service /healthcheck on httpListener {
     resource function get status() returns http:Response {
         http:Response response = new;
+        Context ctx = {variables: {}};
+        ctx.payload = {"status": "UP"};
         response.setPayload({"status": "UP"});
         return response;
     }
