+++ cli-synapse/src/test/resources/ballerina/property1/types.bal
@@ -6,4 +6,5 @@ public type Variables record {|
 
 public type Context record {|
     Variables variables;
+    anydata payload = ();
 |};
