+++ cli-synapse/src/test/resources/ballerina/property2/types.bal
@@ -7,4 +7,5 @@ public type Variables record {|
 
 public type Context record {|
     Variables variables;
+    anydata payload = ();
 |};
