+++ cli-synapse/src/test/resources/ballerina/property5/types.bal
@@ -7,4 +7,5 @@ public type Variables record {|
 
 public type Context record {|
     Variables variables;
+    anydata payload = ();
 |};
