+++ cli-synapse/src/test/resources/ballerina/sequence3/main.bal
@@ -4,16 +4,18 @@ public listener http:Listener httpListener = new (8080);
 
 service /HelloWorld on httpListener {
     resource function get status/[string name]/[string id](string q) returns http:Response {
-        Context ctx = {variables: {}};
         http:Response response = new;
+        Context ctx = {variables: {}};
+        ctx.payload = {"Hello": "World"};
         response.setPayload({"Hello": "World"});
         foo(ctx);
         return response;
     }
 
     resource function get id1/[string id]() returns http:Response {
-        Context ctx = {variables: {}};
         http:Response response = new;
+        Context ctx = {variables: {}};
+        ctx.payload = {"Hello": "World"};
         response.setPayload({"Hello": "World"});
         bar(ctx);
         return response;
