+++ cli-synapse/src/test/resources/ballerina/sequence4/types.bal
@@ -1,7 +1,4 @@
 public type Variables record {|
-    string prop1?;
-    int r1only?;
-    boolean r2only?;
 |};
 
 public type Context record {|
