+++ cli-synapse/src/test/resources/ballerina/sequence5/types.bal
@@ -4,4 +4,5 @@ public type Variables record {|
 
 public type Context record {|
     Variables variables;
+    anydata payload = ();
 |};
