+++ cli-synapse/src/test/resources/ballerina/sequence8/functions.bal
@@ -1,9 +1,10 @@
 import ballerina/http;
 
-function foo(http:Response response) {
+function foo(Context ctx, http:Response response) {
+    ctx.payload = {"id": "002"};
     response.setPayload({"id": "002"});
 }
 
-function wrapper(http:Response response) {
-    foo(response);
+function wrapper(Context ctx, http:Response response) {
+    foo(ctx, response);
 }
