+++ cli-synapse/src/test/resources/ballerina/sequence8/main.bal
@@ -5,6 +5,7 @@ public listener http:Listener httpListener = new (8080);
 service /HelloWorld on httpListener {
     resource function get payload() {
         http:Response response = new;
-        foo(response);
+        Context ctx = {variables: {}};
+        foo(ctx, response);
     }
 }
