+++ cli-synapse/src/test/resources/ballerina/sequence9/types.bal
@@ -5,4 +5,5 @@ public type Variables record {|
 
 public type Context record {|
     Variables variables;
+    anydata payload = ();
 |};
